add utility class for shared functions across modules
