+++ wear/src/main/java/io/github/romantsisyk/wear/presentation/WearMainScreen.kt
@@ -1,5 +1,6 @@
 package io.github.romantsisyk.wear.presentation
 
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -15,11 +16,9 @@ import androidx.compose.ui.unit.dp
 import androidx.wear.compose.foundation.lazy.ScalingLazyColumn
 import androidx.wear.compose.material.CircularProgressIndicator
 import androidx.wear.compose.material.Text
+import io.github.romantsisyk.common.core.generateSunData
 import io.github.romantsisyk.common.presentation.viewmodel.SunriseSunsetViewModel
 import io.github.romantsisyk.common.presentation.viewmodel.UiState
-import java.time.ZonedDateTime
-import java.time.format.DateTimeFormatter
-import java.util.Locale
 
 @Composable
 fun WearMainScreen(viewModel: SunriseSunsetViewModel) {
@@ -36,28 +35,10 @@ fun WearMainScreen(viewModel: SunriseSunsetViewModel) {
             }
 
             is UiState.Success -> {
-                val data = (uiState as UiState.Success).data
-
-                val sunData = listOf(
-                    "☀️" to "Sunrise" to data.sunrise,
-                    "🌙" to "Sunset" to data.sunset,
-                    "🌞" to "Solar Noon" to data.solarNoon,
-                    "⏳" to "Day Length" to data.dayLength,
-                    "🌅" to "Civil Twilight Begin" to data.civilTwilightBegin,
-                    "🌆" to "Civil Twilight End" to data.civilTwilightEnd,
-                    "🌌" to "Nautical Twilight Begin" to data.nauticalTwilightBegin,
-                    "🌉" to "Nautical Twilight End" to data.nauticalTwilightEnd,
-                    "🌠" to "Astronomical Twilight Begin" to data.astronomicalTwilightBegin,
-                    "🌌" to "Astronomical Twilight End" to data.astronomicalTwilightEnd
-                )
-
-                sunData.forEach { (iconLabel, value) ->
+                val sunData = generateSunData((uiState as UiState.Success).data)
+                sunData.forEach { (icon, label, value) ->
                     item {
-                        SunriseSunsetRow(
-                            icon = iconLabel.first,
-                            label = iconLabel.second,
-                            value = value
-                        )
+                        SunriseSunsetRow(icon = icon, label = label, value = value)
                     }
                 }
             }
@@ -79,17 +60,10 @@ fun SunriseSunsetRow(icon: String, label: String, value: String) {
         verticalAlignment = Alignment.CenterVertically
     ) {
         Text(icon)
-        Spacer(modifier = Modifier.width(8.dp))
-        Text("$label: ${value.formatDate()}")
-    }
-}
-
-
-fun String.formatDate(): String {
-    return try {
-        val parsedDate = ZonedDateTime.parse(this)
-        parsedDate.format(DateTimeFormatter.ofPattern("hh:mm a", Locale.getDefault()))
-    } catch (e: Exception) {
-        this
+        Spacer(modifier = Modifier.width(4.dp))
+        Column {
+            Text("$label:")
+            Text(value)
+        }
     }
 }
